shp-01, refactor for PackageRabbitMqMessages generator.
