--- webservices/src/main/java/com/javaschool/webservices/model/ShippingRabbitRPCMessage.java
@@ -1,18 +1,0 @@
-package com.javaschool.webservices.model;
-
-public class ShippingRabbitRPCMessage {
-
-	private String type;
-	
-	public ShippingRabbitRPCMessage(String type) {
-		this.type = type;
-	}
-
-	public String getType() {
-		return type;
-	}
-
-	public void setType(String type) {
-		this.type = type;
-	}
-}
